move externalIdsUrlFormatters to entity_links lib
